feat: refactor for better testability
