+++ src/data/data-providers/google-sheets/credential.ts
@@ -1,6 +1,6 @@
 import { DateTime } from "luxon";
 import { Authenticator, RegisteredAuthenticator } from "../../../types/entity";
-import { ColumnConstraints, RowData } from "../../../types/row";
+import { ColumnConstraints, RowData } from "../../../types/table";
 import { rowToUser } from "./user";
 
 export const CREDENTIAL_SHEET_NAME = "credentials";
