+++ src/data/data-providers/google-sheets/invite.ts
@@ -1,6 +1,6 @@
 import { DateTime } from "luxon";
 import { Invite } from "../../../types/entity";
-import { ColumnConstraints, RowData } from "../../../types/row";
+import { ColumnConstraints, RowData } from "../../../types/table";
 import { rowToUser, userToRow } from "./user";
 
 // invite
