+++ src/data/data-providers/google-sheets/user.ts
@@ -1,6 +1,6 @@
 import { DateTime } from "luxon";
 import { User } from "../../../types/entity";
-import { ColumnConstraints, RowData } from "../../../types/row";
+import { ColumnConstraints, RowData } from "../../../types/table";
 
 export const USER_SHEET_NAME = "users";
 export const USER_CONSTRAINTS: ColumnConstraints = {
